+++ src/main/java/racingcar/domain/Car.java
@@ -1,7 +1,7 @@
 package racingcar.domain;
 
 import java.util.Objects;
-import racingcar.util.Utils;
+import racingcar.util.Validators;
 
 public class Car implements Comparable<Car> {
     private static final int MINIMUM_DRIVE_CONDITION = 4;
@@ -12,8 +12,8 @@ public class Car implements Comparable<Car> {
 
     public Car(String name) {
         this.name = name;
-        Utils.validateCarsNameLength(this.name);
-        Utils.validateCarsNameBlank(this.name);
+        Validators.validateCarsNameLength(this.name);
+        Validators.validateCarsNameBlank(this.name);
     }
 
     public void drive(int value) {
